Fall back to the local data.json asset when the initial load fails
Ensures that even if the GhibliService is down forever, users will still be able to load Ghibli artwork.

Since there's no requirement that the data.json matches what we get from the network, removed the unnecessary {"artworks":[]} structure and just made the file return the JSONArray that is actually needed.